# backend/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/maintenance": {
      "post": {
        "operationId": "createMaintenanceRequest",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMaintenanceDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MaintenanceRequest"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "",
        "tags": [
          "Maintenance"
        ]
      },
      "get": {
        "operationId": "getMaintenanceRequests",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "default": 10,
              "type": "number"
            }
          },
          {
            "name": "tenantId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "roomId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "PENDING",
                "IN_PROGRESS",
                "COMPLETED",
                "CANCELLED"
              ]
            }
          },
          {
            "name": "priority",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "LOW",
                "MEDIUM",
                "HIGH",
                "URGENT"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MaintenanceRequest"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "",
        "tags": [
          "Maintenance"
        ]
      }
    },
    "/api/v1/maintenance/{id}": {
      "get": {
        "operationId": "getMaintenanceRequest",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MaintenanceRequest"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "",
        "tags": [
          "Maintenance"
        ]
      },
      "patch": {
        "operationId": "updateMaintenanceRequest",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMaintenanceDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MaintenanceRequest"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "",
        "tags": [
          "Maintenance"
        ]
      }
    },
    "/api/v1/maintenance/{id}/respond": {
      "post": {
        "operationId": "respondMaintenanceRequest",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RespondMaintenanceDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MaintenanceRequest"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "",
        "tags": [
          "Maintenance"
        ]
      }
    },
    "/api/v1/chat/groups": {
      "get": {
        "operationId": "getChatGroups",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ChatGroup"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "",
        "tags": [
          "Chat"
        ]
      }
    },
    "/api/v1/chat/groups/{id}": {
      "get": {
        "operationId": "getChatGroup",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatGroup"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "",
        "tags": [
          "Chat"
        ]
      }
    },
    "/api/v1/chat/groups/{id}/messages": {
      "get": {
        "operationId": "getGroupMessages",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "default": 50,
              "type": "number"
            }
          },
          {
            "name": "before",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Message"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "",
        "tags": [
          "Chat"
        ]
      },
      "post": {
        "operationId": "sendGroupMessage",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendMessageDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "",
        "tags": [
          "Chat"
        ]
      }
    },
    "/api/v1/chat/direct/{userId}": {
      "get": {
        "operationId": "getDirectMessages",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "default": 50,
              "type": "number"
            }
          },
          {
            "name": "before",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Message"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "",
        "tags": [
          "Chat"
        ]
      },
      "post": {
        "operationId": "sendDirectMessage",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendMessageDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "",
        "tags": [
          "Chat"
        ]
      }
    },
...
  },
  "info": {
    "title": "TacoHouse API",
    "description": "A comprehensive rental room management system API for managing buildings, rooms, tenants, billing, and payments.\n\n## Response Format\nAll responses follow a standardized format:\n- **status**: HTTP status code\n- **message**: Human-readable message\n- **data**: Response data (object or array)\n- **pagination**: (Optional) Pagination metadata for list endpoints",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3005",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Enter JWT token"
      }
    },
    "schemas": {
...
      "UserRole": {
        "type": "string",
        "enum": [
          "ADMIN",
          "LANDLORD",
          "TENANT"
        ]
      },
...
      "PriorityType": {
        "type": "string",
        "enum": [
          "LOW",
          "MEDIUM",
          "HIGH",
          "URGENT"
        ]
      },
      "MaintenanceCategory": {
        "type": "string",
        "enum": [
          "PLUMBING",
          "ELECTRICAL",
          "APPLIANCE",
          "FURNITURE",
          "CLEANING",
          "OTHER"
        ]
      },
      "MaintenanceStatus": {
        "type": "string",
        "enum": [
          "PENDING",
          "IN_PROGRESS",
          "COMPLETED",
          "CANCELLED"
        ]
      },
      "MaintenanceRequest": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "tenantId": {
            "type": "string"
          },
          "tenantRole": {
            "allOf": [
              {
                "$ref": "#/components/schemas/UserRole"
              }
            ]
          },
          "roomId": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "priority": {
            "allOf": [
              {
                "$ref": "#/components/schemas/PriorityType"
              }
            ]
          },
          "category": {
            "allOf": [
              {
                "$ref": "#/components/schemas/MaintenanceCategory"
              }
            ]
          },
          "images": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "status": {
            "allOf": [
              {
                "$ref": "#/components/schemas/MaintenanceStatus"
              }
            ]
          },
          "completedAt": {
            "type": "string",
            "nullable": true,
            "format": "date-time"
          },
          "completionNote": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "tenantId",
          "tenantRole",
          "roomId",
          "title",
          "description",
          "priority",
          "category",
          "images",
          "status",
          "completedAt",
          "completionNote",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "MessageType": {
        "type": "string",
        "enum": [
          "TEXT",
          "IMAGE",
          "FILE"
        ]
      },
      "Message": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "senderId": {
            "type": "string"
          },
          "content": {
            "type": "string"
          },
          "messageType": {
            "allOf": [
              {
                "$ref": "#/components/schemas/MessageType"
              }
            ]
          },
          "attachmentUrl": {
            "type": "string",
            "nullable": true
          },
          "attachmentName": {
            "type": "string",
            "nullable": true
          },
          "attachmentSize": {
            "type": "integer",
            "nullable": true,
            "format": "int32"
          },
          "chatGroupId": {
            "type": "string",
            "nullable": true
          },
          "recipientId": {
            "type": "string",
            "nullable": true
          },
          "readAt": {
            "type": "string",
            "nullable": true,
            "format": "date-time"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "senderId",
          "content",
          "messageType",
          "attachmentUrl",
          "attachmentName",
          "attachmentSize",
          "chatGroupId",
          "recipientId",
          "readAt",
          "createdAt",
          "updatedAt"
        ]
      },
      "ChatGroupMember": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "chatGroupId": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          },
          "joinedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "chatGroupId",
          "userId",
          "joinedAt"
        ]
      },
      "ChatGroup": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "buildingId": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "messages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Message"
            }
          },
          "members": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ChatGroupMember"
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "buildingId",
          "name",
          "description",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "CreateMaintenanceDto": {
        "type": "object",
        "properties": {
          "roomId": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "priority": {
            "type": "string",
            "enum": [
              "LOW",
              "MEDIUM",
              "HIGH",
              "URGENT"
            ]
          },
          "category": {
            "type": "string",
            "enum": [
              "PLUMBING",
              "ELECTRICAL",
              "APPLIANCE",
              "FURNITURE",
              "CLEANING",
              "OTHER"
            ]
          },
          "images": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "roomId",
          "title",
          "description",
          "priority",
          "category"
        ]
      },
      "UpdateMaintenanceDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "priority": {
            "type": "string",
            "enum": [
              "LOW",
              "MEDIUM",
              "HIGH",
              "URGENT"
            ]
          },
          "category": {
            "type": "string",
            "enum": [
              "PLUMBING",
              "ELECTRICAL",
              "APPLIANCE",
              "FURNITURE",
              "CLEANING",
              "OTHER"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "IN_PROGRESS",
              "COMPLETED",
              "CANCELLED"
            ]
          },
          "completionNote": {
            "type": "string"
          }
        }
      },
      "RespondMaintenanceDto": {
        "type": "object",
        "properties": {
          "response": {
            "type": "string"
          }
        },
        "required": [
          "response"
        ]
      },
      "SendMessageDto": {
        "type": "object",
        "properties": {
          "content": {
            "type": "string"
          },
          "messageType": {
            "enum": [
              "TEXT",
              "IMAGE",
              "FILE"
            ],
            "type": "string"
          }
        },
        "required": [
          "content"
        ]
      },
...
    }
  }
}